pyproject.toml
[project]
name = "alphabetum"
version = "0.1.0"
description = "Reconstructing Leibniz's Alphabet of Human Thought"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.25.0",
    "openai>=1.12.0",
    "pyyaml>=6.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "networkx>=3.2",
    "sympy>=1.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
viz = [
    "matplotlib>=3.8.0",
    "graphviz>=0.20",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
